# apps/backend-nest/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Vben Backend API",
    "description": "基于 NestJS + Zod 构建的现代化后端 API 服务\n\n## 架构特点\n- 使用 Zod 进行数据验证和类型定义\n- 统一的三段式返回格式 (code, message, data)\n- 自动生成的 OpenAPI 文档\n- 类型安全的 API 接口",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.example.com",
      "description": "生产环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 认证令牌，格式：Bearer <token>"
      }
    }
  },
  "paths": {
...
    "/api/system/menu": {
      "post": {
        "summary": "创建菜单",
        "description": "创建新的菜单项",
        "tags": [
          "菜单管理"
        ],
        "operationId": "MenuController_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "description": "菜单名称（唯一）"
                  },
                  "path": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 200,
                    "description": "路由路径（唯一，可为空）"
                  },
                  "component": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 200,
                    "description": "组件路径"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "catalog",
                      "menu",
                      "button",
                      "embedded",
                      "link"
                    ],
                    "description": "菜单类型：catalog-目录，menu-菜单，button-按钮，embedded-内嵌，link-外链"
                  },
                  "authCode": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 100,
                    "description": "权限标识码"
                  },
                  "pid": {
                    "type": "string",
                    "nullable": true,
                    "description": "父级菜单ID，支持树形结构"
                  },
                  "status": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1,
                    "default": 1,
                    "description": "菜单状态：0-禁用，1-启用"
                  },
                  "meta": {
                    "type": "object",
                    "nullable": true,
                    "properties": {
                      "title": {
                        "type": "string",
                        "description": "菜单标题"
                      },
                      "icon": {
                        "type": "string",
                        "description": "菜单图标"
                      },
                      "hideInMenu": {
                        "type": "boolean",
                        "description": "是否在菜单中隐藏"
                      },
                      "hideInBreadcrumb": {
                        "type": "boolean",
                        "description": "是否在面包屑中隐藏"
                      },
                      "hideChildrenInMenu": {
                        "type": "boolean",
                        "description": "是否隐藏子菜单"
                      },
                      "orderNo": {
                        "type": "number",
                        "description": "排序号"
                      },
                      "frameSrc": {
                        "type": "string",
                        "description": "内嵌框架地址"
                      },
                      "externalLink": {
                        "type": "string",
                        "description": "外部链接地址"
                      },
                      "ignoreKeepAlive": {
                        "type": "boolean",
                        "description": "是否忽略缓存"
                      },
                      "affix": {
                        "type": "boolean",
                        "description": "是否固定标签页"
                      },
                      "carryParam": {
                        "type": "boolean",
                        "description": "是否携带参数"
                      },
                      "single": {
                        "type": "boolean",
                        "description": "是否单独显示"
                      },
                      "currentActiveMenu": {
                        "type": "string",
                        "description": "当前激活菜单"
                      },
                      "hideTab": {
                        "type": "boolean",
                        "description": "是否隐藏标签页"
                      },
                      "hideMenu": {
                        "type": "boolean",
                        "description": "是否隐藏菜单"
                      },
                      "isLink": {
                        "type": "boolean",
                        "description": "是否为外链"
                      },
                      "ignoreRoute": {
                        "type": "boolean",
                        "description": "是否忽略路由"
                      },
                      "hidePathForChildren": {
                        "type": "boolean",
                        "description": "是否为子级隐藏路径"
                      }
                    },
                    "description": "菜单元数据（JSON格式存储图标、标题等信息）"
                  }
                },
                "required": [
                  "name",
                  "type"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "菜单唯一标识"
                    },
                    "name": {
                      "type": "string",
                      "maxLength": 100,
                      "description": "菜单名称（唯一）"
                    },
                    "path": {
                      "type": "string",
                      "nullable": true,
                      "maxLength": 200,
                      "description": "路由路径（唯一，可为空）"
                    },
                    "component": {
                      "type": "string",
                      "nullable": true,
                      "maxLength": 200,
                      "description": "组件路径"
                    },
                    "type": {
                      "type": "string",
                      "maxLength": 20,
                      "description": "菜单类型：catalog-目录，menu-菜单，button-按钮，embedded-内嵌，link-外链"
                    },
                    "authCode": {
                      "type": "string",
                      "nullable": true,
                      "maxLength": 100,
                      "description": "权限标识码"
                    },
                    "pid": {
                      "type": "string",
                      "nullable": true,
                      "description": "父级菜单ID，支持树形结构"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1,
                      "description": "菜单状态：0-禁用，1-启用"
                    },
                    "meta": {
                      "nullable": true,
                      "description": "菜单元数据（JSON格式存储图标、标题等信息）"
                    },
                    "createTime": {
                      "type": "string",
                      "format": "date",
                      "description": "创建时间"
                    },
                    "updateTime": {
                      "type": "string",
                      "format": "date",
                      "description": "更新时间"
                    },
                    "parentId": {
                      "type": "string",
                      "nullable": true,
                      "description": "父级菜单ID"
                    },
                    "childrenIds": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "子级菜单ID列表"
                    },
                    "rolePermissions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "关联记录唯一标识"
                          },
                          "roleId": {
                            "type": "string",
                            "description": "角色ID"
                          },
                          "menuId": {
                            "type": "string",
                            "description": "菜单ID"
                          },
                          "createTime": {
                            "type": "string",
                            "format": "date",
                            "description": "创建时间"
                          },
                          "role": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "description": "角色唯一标识"
                              },
                              "name": {
                                "type": "string",
                                "maxLength": 100,
                                "description": "角色名称（唯一）"
                              },
                              "status": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 1,
                                "description": "角色状态：0-禁用，1-启用"
                              },
                              "remark": {
                                "type": "string",
                                "nullable": true,
                                "description": "备注信息"
                              },
                              "createTime": {
                                "type": "string",
                                "format": "date",
                                "description": "创建时间"
                              },
                              "updateTime": {
                                "type": "string",
                                "format": "date",
                                "description": "更新时间"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "status",
                              "createTime",
                              "updateTime"
                            ],
                            "description": "关联的角色信息"
                          }
                        },
                        "required": [
                          "id",
                          "roleId",
                          "menuId",
                          "createTime"
                        ]
                      },
                      "description": "角色权限关联"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "type",
                    "status",
                    "createTime",
                    "updateTime"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "请求参数验证失败"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/system/menu/list": {
      "get": {
        "summary": "获取菜单列表",
        "description": "获取菜单列表，支持筛选条件",
        "tags": [
          "菜单管理"
        ],
        "operationId": "MenuController_findAll",
        "parameters": [
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "default": 1,
              "description": "页码，从1开始"
            },
            "required": false,
            "description": "页码，从1开始",
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "description": "每页数量，最大100条"
            },
            "required": false,
            "description": "每页数量，最大100条",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "按菜单名称筛选"
            },
            "required": false,
            "description": "按菜单名称筛选",
            "name": "name",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "catalog",
                "menu",
                "button",
                "embedded",
                "link"
              ],
              "description": "按菜单类型筛选"
            },
            "required": false,
            "description": "按菜单类型筛选",
            "name": "type",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1,
              "description": "按菜单状态筛选：0-禁用，1-启用"
            },
            "required": false,
            "description": "按菜单状态筛选：0-禁用，1-启用",
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "按父级菜单ID筛选"
            },
            "required": false,
            "description": "按父级菜单ID筛选",
            "name": "pid",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "菜单唯一标识"
                    },
                    "name": {
                      "type": "string",
                      "maxLength": 100,
                      "description": "菜单名称（唯一）"
                    },
                    "path": {
                      "type": "string",
                      "nullable": true,
                      "maxLength": 200,
                      "description": "路由路径（唯一，可为空）"
                    },
                    "component": {
                      "type": "string",
                      "nullable": true,
                      "maxLength": 200,
                      "description": "组件路径"
                    },
                    "type": {
                      "type": "string",
                      "maxLength": 20,
                      "description": "菜单类型：catalog-目录，menu-菜单，button-按钮，embedded-内嵌，link-外链"
                    },
                    "authCode": {
                      "type": "string",
                      "nullable": true,
                      "maxLength": 100,
                      "description": "权限标识码"
                    },
                    "pid": {
                      "type": "string",
                      "nullable": true,
                      "description": "父级菜单ID，支持树形结构"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1,
                      "description": "菜单状态：0-禁用，1-启用"
                    },
                    "meta": {
                      "nullable": true,
                      "description": "菜单元数据（JSON格式存储图标、标题等信息）"
                    },
                    "createTime": {
                      "type": "string",
                      "format": "date",
                      "description": "创建时间"
                    },
                    "updateTime": {
                      "type": "string",
                      "format": "date",
                      "description": "更新时间"
                    },
                    "parentId": {
                      "type": "string",
                      "nullable": true,
                      "description": "父级菜单ID"
                    },
                    "childrenIds": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "子级菜单ID列表"
                    },
                    "rolePermissions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "关联记录唯一标识"
                          },
                          "roleId": {
                            "type": "string",
                            "description": "角色ID"
                          },
                          "menuId": {
                            "type": "string",
                            "description": "菜单ID"
                          },
                          "createTime": {
                            "type": "string",
                            "format": "date",
                            "description": "创建时间"
                          },
                          "role": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "description": "角色唯一标识"
                              },
                              "name": {
                                "type": "string",
                                "maxLength": 100,
                                "description": "角色名称（唯一）"
                              },
                              "status": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 1,
                                "description": "角色状态：0-禁用，1-启用"
                              },
                              "remark": {
                                "type": "string",
                                "nullable": true,
                                "description": "备注信息"
                              },
                              "createTime": {
                                "type": "string",
                                "format": "date",
                                "description": "创建时间"
                              },
                              "updateTime": {
                                "type": "string",
                                "format": "date",
                                "description": "更新时间"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "status",
                              "createTime",
                              "updateTime"
                            ],
                            "description": "关联的角色信息"
                          }
                        },
                        "required": [
                          "id",
                          "roleId",
                          "menuId",
                          "createTime"
                        ]
                      },
                      "description": "角色权限关联"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "type",
                    "status",
                    "createTime",
                    "updateTime"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "请求参数验证失败"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/system/menu/name-exists": {
      "get": {
        "summary": "检查菜单名称是否存在",
        "description": "检查菜单名称是否已存在",
        "tags": [
          "菜单管理"
        ],
        "operationId": "MenuController_isNameExists",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "菜单名称（唯一）"
            },
            "required": true,
            "description": "菜单名称（唯一）",
            "name": "name",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "菜单ID，用于排除自身"
            },
            "required": false,
            "description": "菜单ID，用于排除自身",
            "name": "id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "请求参数验证失败"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/system/menu/path-exists": {
      "get": {
        "summary": "检查菜单路径是否存在",
        "description": "检查菜单路径是否已存在",
        "tags": [
          "菜单管理"
        ],
        "operationId": "MenuController_isPathExists",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "路由路径（唯一，可为空）"
            },
            "required": true,
            "description": "路由路径（唯一，可为空）",
            "name": "path",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "菜单ID，用于排除自身"
            },
            "required": false,
            "description": "菜单ID，用于排除自身",
            "name": "id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "请求参数验证失败"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}